[DEV-729] Refactoring
