+++ src/payment/models/route/dto/route-history.dto.ts
@@ -1,5 +1,11 @@
 import { AmlCheck } from '../../buy-crypto/enums/aml-check.enum';
 
+export enum HistoryTransactionType {
+  BUY = 'Buy',
+  SELL = 'Sell',
+  CRYPTO = 'Crypto',
+}
+
 export class RouteHistoryDto {
   inputAmount: number;
   inputAsset: string;
@@ -9,4 +15,8 @@ export class RouteHistoryDto {
   date: Date;
   amlCheck: AmlCheck;
   isComplete: boolean;
+}
+
+export class TypedRouteHistoryDto extends RouteHistoryDto {
+  type: HistoryTransactionType;
 }
